created homepage.jsx and favpage.jsx, refactored app.jsx and moved actions to favpage.jsx
